fix(timeline): Incorporate Copilot review feedback
