+++ packages/jaeger-ui/src/components/TracePage/TraceTimelineViewer/VirtualizedTraceView.tsx
@@ -25,9 +25,9 @@ import {
   isKindClient,
   isKindProducer,
   spanContainsErredSpan,
-  hasGenAIWarning,
   spanContainsWarningSpan,
 } from './utils';
+import { hasGenAIWarning } from '../../../utils/genai';
 import { Accessors } from '../ScrollManager';
 import { parseUiFind, TExtractUiFindFromStateReturn } from '../../common/UiFindInput';
 import getLinks from '../../../model/link-patterns';
@@ -368,8 +368,9 @@ export const VirtualizedTraceViewImpl = React.memo(function VirtualizedTraceView
       const isSelected = selectedSpanID === spanID;
       const hasOwnError = isErrorSpan(span);
       const hasChildError = isCollapsed && spanContainsErredSpan(spans, spanIndex);
-      const hasOwnWarning = hasGenAIWarning(span);
-      const hasChildWarning = isCollapsed && spanContainsWarningSpan(spans, spanIndex);
+      const hasOwnWarning = !hasOwnError && hasGenAIWarning(span);
+      const hasChildWarning =
+        isCollapsed && !hasOwnError && !hasChildError && spanContainsWarningSpan(spans, spanIndex);
       const hasPrunedChildren =
         prunedServices.size > 0 &&
         span.childSpans.some(child => prunedServices.has(child.resource.serviceName));
